Extract audio signal-chain diagram into its own component
Keeps QualityDetailsBtn focused on the trigger button and popover shell.

- Move the signal-chain diagram into a dedicated StreamDetailsChain component
- Add a useStreamQuality composable for the shared quality-tier logic
- QualityDetailsBtn now just renders the trigger and hosts the diagram
